+++ include/Nazara/Network/ENetPeer.hpp
@@ -58,6 +58,7 @@ namespace Nz
 			inline ENetPeerState GetState() const;
 			inline UInt64 GetTotalByteReceived() const;
 			inline UInt64 GetTotalByteSent() const;
+			inline UInt64 GetTotalPacketReceived() const;
 			inline UInt64 GetTotalPacketLost() const;
 			inline UInt64 GetTotalPacketSent() const;
 
@@ -241,6 +242,7 @@ namespace Nz
 			UInt32                                m_windowSize;
 			UInt64                                m_totalByteReceived;
 			UInt64                                m_totalByteSent;
+			UInt64                                m_totalPacketReceived;
 			UInt64                                m_totalPacketLost;
 			UInt64                                m_totalPacketSent;
 			bool                                  m_isSimulationEnabled;